# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kenya Administrative Boundaries API(KenyaAdminAPI) - Documentation",
        "description": "The Kenya Administrative Boundaries API (KenAdminAPI) provides a comprehensive and easy-to-use interface for accessing detailed information about Kenya's administrative divisions, including Counties, Constituencies, and Wards. This API is designed to help developers integrate location-based data into their applications, supporting a wide range of use cases from geographical analysis to service delivery.",
        "contact": {
            "email": "sosmongare@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "https://127.0.0.1:8000",
            "description": "Local secure server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/constituencies": {
            "get": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Get list of constituencies",
                "operationId": "82e23da030eb11f9a17938c0114c75b5",
                "responses": {
                    "200": {
                        "description": "Constituencies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Constituency"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Create a new constituency",
                "operationId": "d112701fedbfa4580123aa277f73d94b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "county_id",
                                    "constituency_name"
                                ],
                                "properties": {
                                    "county_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "constituency_name": {
                                        "type": "string",
                                        "example": "Westlands"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Constituency created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Constituency"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/constituencies/{id}": {
            "get": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Get a constituency by ID",
                "operationId": "582332ca73c91da2140d75e45c5689ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Constituency retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Constituency"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Update a constituency",
                "operationId": "ce3bd866f4a6489e7481d8002a194b6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "county_id",
                                    "constituency_name"
                                ],
                                "properties": {
                                    "county_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "constituency_name": {
                                        "type": "string",
                                        "example": "Westlands"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Constituency updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Constituency"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Delete a constituency",
                "operationId": "751381a64026663cf053bb8ff41ff641",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Constituency deleted successfully"
                    }
                }
            }
        },
        "/api/constituencies/search/{query}": {
            "get": {
                "tags": [
                    "Constituencies"
                ],
                "summary": "Search for constituencies",
                "operationId": "f4c29d61169fc61229d82667d5e02f6f",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Constituency"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No results found for the given query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Constituency": {
                "title": "Constituency",
                "description": "Constituency model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the constituency",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "county_id": {
                        "title": "County ID",
                        "description": "ID of the county the constituency belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "constituency_name": {
                        "title": "Constituency Name",
                        "description": "Name of the constituency",
                        "type": "string",
                        "example": "Westlands"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Constituency"
                }
            },
...
        }
    },
...
}